refactor: move EmbeddedMcpServer to provider
